+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `PathUtils` removed, `PathPredicates` rework
+- Line extension: empty string is permitted
 
 ---
 ## [0.0.4] - 2025-09-27